python-layer/requirements.txt
redis
python-dotenv
openai
boto3
requests